+++ src/pynwb/tests/test_constructors.py
@@ -7,8 +7,8 @@ from ndx_microscopy.testing import (
     mock_ExcitationLightPath,
     mock_Microscope,
     mock_Segmentation,
-    mock_Segmentation2D,
-    mock_Segmentation3D,
+    mock_PlanarSegmentation,
+    mock_VolumetricSegmentation,
     mock_SegmentationContainer,
     mock_PlanarImagingSpace,
     mock_PlanarMicroscopySeries,
@@ -24,8 +24,8 @@ from ndx_microscopy.testing import (
 )
 from ndx_microscopy import (
     Segmentation,
-    Segmentation2D,
-    Segmentation3D,
+    PlanarSegmentation,
+    VolumetricSegmentation,
     PlanarImagingSpace,
     VolumetricImagingSpace,
     IlluminationPattern,
@@ -125,27 +125,27 @@ def test_constructor_segmentation():
     assert isinstance(segmentation, Segmentation)
 
 
-def test_constructor_segmentation_2D():
-    """Test constructor for Segmentation2D class."""
+def test_constructor_planar_segmentation():
+    """Test constructor for PlanarSegmentation class."""
     planar_imaging_space = mock_PlanarImagingSpace()
-    segmentation = mock_Segmentation2D(planar_imaging_space=planar_imaging_space)
-    assert segmentation.description == "A mock instance of a Segmentation2D type to be used for rapid testing."
+    segmentation = mock_PlanarSegmentation(planar_imaging_space=planar_imaging_space)
+    assert segmentation.description == "A mock instance of a PlanarSegmentation type to be used for rapid testing."
     assert len(segmentation.id) == 5  # Default number_of_rois
     assert "image_mask" in segmentation.colnames
     assert isinstance(segmentation.planar_imaging_space, PlanarImagingSpace)
-    assert isinstance(segmentation, Segmentation2D)
+    assert isinstance(segmentation, PlanarSegmentation)
     assert isinstance(segmentation, Segmentation)  # Test inheritance
 
 
 def test_constructor_volumetric_segmentation():
-    """Test constructor for Segmentation3D class."""
+    """Test constructor for VolumetricSegmentation class."""
     volumetric_imaging_space = mock_VolumetricImagingSpace()
-    segmentation = mock_Segmentation3D(volumetric_imaging_space=volumetric_imaging_space)
-    assert segmentation.description == "A mock instance of a Segmentation3D type to be used for rapid testing."
+    segmentation = mock_VolumetricSegmentation(volumetric_imaging_space=volumetric_imaging_space)
+    assert segmentation.description == "A mock instance of a VolumetricSegmentation type to be used for rapid testing."
     assert len(segmentation.id) == 5  # Default number_of_rois
     assert "image_mask" in segmentation.colnames
     assert isinstance(segmentation.volumetric_imaging_space, VolumetricImagingSpace)
-    assert isinstance(segmentation, Segmentation3D)
+    assert isinstance(segmentation, VolumetricSegmentation)
     assert isinstance(segmentation, Segmentation)  # Test inheritance
 
 
@@ -154,8 +154,8 @@ def test_constructor_segmentation_container():
     container = mock_SegmentationContainer()
     assert len(container.segmentations) == 2  # Default includes both planar and volumetric
     segmentation_names = [seg_name for seg_name in container.segmentations]
-    assert isinstance(container.segmentations[segmentation_names[0]], Segmentation2D)
-    assert isinstance(container.segmentations[segmentation_names[1]], Segmentation3D)
+    assert isinstance(container.segmentations[segmentation_names[0]], PlanarSegmentation)
+    assert isinstance(container.segmentations[segmentation_names[1]], VolumetricSegmentation)
 
 
 def test_constructor_planar_microscopy_series():
@@ -216,7 +216,7 @@ def test_constructor_volumetric_microscopy_series():
 def test_constructor_microscopy_response_series():
     number_of_rois = 10
     planar_imaging_space = mock_PlanarImagingSpace()
-    segmentation = mock_Segmentation2D(planar_imaging_space=planar_imaging_space, number_of_rois=number_of_rois)
+    segmentation = mock_PlanarSegmentation(planar_imaging_space=planar_imaging_space, number_of_rois=number_of_rois)
 
     rois = segmentation.create_roi_table_region(
         description="test region",
@@ -233,7 +233,7 @@ def test_constructor_microscopy_response_series():
 def test_constructor_microscopy_response_series_container():
     number_of_rois = 10
     planar_imaging_space = mock_PlanarImagingSpace()
-    segmentation = mock_Segmentation2D(planar_imaging_space=planar_imaging_space, number_of_rois=number_of_rois)
+    segmentation = mock_PlanarSegmentation(planar_imaging_space=planar_imaging_space, number_of_rois=number_of_rois)
 
     rois = segmentation.create_roi_table_region(
         description="test region",